Fixed up the timezone mayhem
If a server wasn't in the local time zone of the rinks (because servers should rightfully be UTC rather than the local timezone), the times of the games were all messed up.  Rinks and leagues as implemented by the parsers are now anchored to the timezone they reside in.  Game objects will still contain UTC times and the ScheduleRipper is responsible for figuring out what timezone the calendar is set for and making the conversions as they're found via search