Added some hashing action to ensure that game numbers are unique for multiple teams
Ice Angeles doesn't have a unique game number like most other leagues do, so we have to generate one on our own.  Started off as just a count as a test, and that works for a single team, but having multiple teams means that you end up with multiple game 3's (for example).  Now we're hashing the team name and the number to generate a repeatable unique identifier for each game